refactor worker code structure
